adjust DumpParentEntry and directory tree's print format
